Decode location coordinates

+++ TilesApp/TilesApp/TilesApp/Models/Location.cs
@@ -6,10 +6,10 @@ namespace TilesApp.Models
 {
     public class Location
     {
-        public int place_id { get; set; }
+        public long place_id { get; set; }
         public string licence { get; set; }
         public string osm_type { get; set; }
-        public int osm_id { get; set; }
+        public long osm_id { get; set; }
         public string lat { get; set; }
         public string lon { get; set; }
         public string display_name { get; set; }
